Rewrite cache-flush.c.
When switching from gcc5 to gcc7, we failed cache-flush.
After rewriting it, we passed it, I think it has something to do with
the operand evaluation ordering.
To be specefic, these two sentence:
* arr[fast_rand() & lengthMod] = fast_rand();
* nemu_assert(arr[fast_rand() & lengthMod] == fast_rand());

* The first fast_rand does not necessarily evaluate before the second
one, there may be some undefined or implementation defined behavior
here. I do not know for sure.
Rewrite it to be:
        int idx = fast_rand() & lengthMod;
        int result = fast_rand();
Now the first fast_rand will execute before the second one as fast_rand
has side effects, according to program order, the first one will be
evaluated first.